backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-mail==0.9.1
pymongo==4.6.1
python-dotenv==1.0.0
itsdangerous==2.1.2
requests==2.31.0
numpy==1.26.3
scikit-learn==1.4.0
transformers==4.36.2
torch==2.1.2
python-docx==1.1.0
PyPDF2==3.0.1
beautifulsoup4==4.12.2
nltk==3.8.1
spacy==3.7.2
pdfminer.six==20221105
chardet==5.2.0
